Update action buttons to bar and disabled vuex store
Extracted the actions into a component with multiple
content options. This bar is used in all components
that need a back button, title, edit button and/or remove
button. All calls and references to vuex are (temporarily)
commented. First we want function, then speed.
Might be used for single (active/current) items, not
for caching for now.