Start map input and theme and separate locationables component
Started working on the map input. When placing a
marker, the coordinates and zoom are emitted up.
Added theme for the maps using ol-ext. Not perfect
but it'll do at least for now.

The locationables will be handled in separate components.
Higher components need not to worry about poi vs
location anymore in this case, all is encapsulated.