Multiple locationables per section and draggable within the section
Multiple locationables can now be added in one section
and the user can decide where in a section a locationable is
placed by dragging it in the correct position. Controllers
are updated, models as well. Migrations are updated and
created and the representation for locations and pois
resources are updated.